requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
google-generativeai==0.3.2
python-dotenv==1.0.0
numpy==1.24.3
aiofiles==0.23.2
pydantic==2.5.0
streamlit==1.28.1
requests==2.31.0
